♻️ Move read_updater_blacklist to update module
