+++ src/config.rs
@@ -1,12 +1,9 @@
 use std::{
-    collections::HashSet,
     env,
-    fs::File,
-    io::{self, BufRead, BufReader},
     path::{Path, PathBuf},
 };
 
-use tracing::{error, warn};
+use tracing::warn;
 
 pub const CARGO_PKG_NAME: &str = env!("CARGO_PKG_NAME");
 pub const CARGO_PKG_VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -64,40 +61,6 @@ impl AppConfig {
 
     pub fn cache_db_path(&self) -> &Path {
         &self.cache_db_path
-    }
-
-    // FIXME Create this field in AppConfig or getter
-    const UPDATER_BLACKLIST_FILE: &str = "updaterblacklist.txt";
-
-    // FIXME `AppConfig` should not read the File or perform any actions, move this method to outside
-    /// Reads `updaterblacklist.txt` and returns blacklisted mod paths.
-    pub fn read_updater_blacklist(&self) -> io::Result<HashSet<String>> {
-        let path = self.mods_dir().join(Self::UPDATER_BLACKLIST_FILE);
-        let mut blacklist = HashSet::new();
-
-        let mut file = match File::open(&path) {
-            Ok(value) => value,
-            // NOTE It's safe to ignore NotFound here, as this file is optional.
-            Err(err) if err.kind().eq(&io::ErrorKind::NotFound) => return Ok(blacklist),
-            Err(err) => {
-                error!(?err, "failed to open blacklist file");
-                return Err(err);
-            }
-        };
-
-        // NOTE The default 8KiB buffer is overkill for small text files.
-        let reader = BufReader::with_capacity(1024, &mut file);
-
-        for line in reader.lines() {
-            let line = line.inspect_err(|err| error!(?err, "failed to read line in blacklist"))?;
-            let line = line.trim();
-            if !line.starts_with('#') && !line.is_empty() {
-                warn!("'{}' will be excluded from updates", line);
-                blacklist.insert(line.to_string());
-            }
-        }
-
-        Ok(blacklist)
     }
 }
